refactor(calculator): Number 예외처리
